Incorporate idempotence checks to file registry

+++ services/ifrs/src/ifrs/core/file_registry.py
@@ -24,8 +24,13 @@ from ghga_service_commons.utils.multinode_storage import ObjectStorages
 
 from ifrs.config import Config
 from ifrs.core import models
+from ifrs.core.utils import assert_record_is_new, make_record_from_update
 from ifrs.ports.inbound.file_registry import FileRegistryPort
-from ifrs.ports.outbound.dao import FileMetadataDaoPort, ResourceNotFoundError
+from ifrs.ports.outbound.dao import (
+    FileMetadataDaoPort,
+    OutboxDaoCollectionPort,
+    ResourceNotFoundError,
+)
 from ifrs.ports.outbound.event_pub import EventPublisherPort
 
 log = logging.getLogger(__name__)
@@ -34,10 +39,11 @@ log = logging.getLogger(__name__)
 class FileRegistry(FileRegistryPort):
     """A service that manages a registry files stored on a permanent object storage."""
 
-    def __init__(
+    def __init__(  # noqa: PLR0913
         self,
         *,
         file_metadata_dao: FileMetadataDaoPort,
+        outbox_dao_collection: OutboxDaoCollectionPort,
         event_publisher: EventPublisherPort,
         object_storages: ObjectStorages,
         config: Config,
@@ -45,6 +51,7 @@ class FileRegistry(FileRegistryPort):
         """Initialize with essential config params and outbound adapters."""
         self._event_publisher = event_publisher
         self._file_metadata_dao = file_metadata_dao
+        self._outbox_dao_collection = outbox_dao_collection
         self._object_storages = object_storages
         self._config = config
 
@@ -96,6 +103,8 @@ class FileRegistry(FileRegistryPort):
         Raises:
             self.FileContentNotInStagingError:
                 When the file content is not present in the storage staging.
+            ValueError:
+                When the configuration for the storage alias is not found.
         """
         storage_alias = file_without_object_id.storage_alias
 
@@ -308,12 +317,19 @@ class FileRegistry(FileRegistryPort):
             update:
                 The NonStagedFileRequested event to upsert.
         """
-        await self.stage_registered_file(
-            file_id=resource_id,
-            decrypted_sha256=update.decrypted_sha256,
-            outbox_object_id=update.target_object_id,
-            outbox_bucket_id=update.target_bucket_id,
-        )
+        dao = self._outbox_dao_collection.get_nonstaged_file_requested_dao()
+
+        record = make_record_from_update(models.NonStagedFileRequestedRecord, update)
+        if await assert_record_is_new(
+            dao=dao, resource_id=resource_id, update=update, record=record
+        ):
+            await self.stage_registered_file(
+                file_id=resource_id,
+                decrypted_sha256=update.decrypted_sha256,
+                outbox_object_id=update.target_object_id,
+                outbox_bucket_id=update.target_bucket_id,
+            )
+            await dao.insert(record)
 
     async def upsert_file_deletion_requested(
         self, *, resource_id: str, update: event_schemas.FileDeletionRequested
@@ -327,7 +343,14 @@ class FileRegistry(FileRegistryPort):
             update:
                 The FileDeletionRequested event to upsert.
         """
-        await self.delete_file(file_id=resource_id)
+        dao = self._outbox_dao_collection.get_file_deletion_requested_dao()
+
+        record = make_record_from_update(models.FileDeletionRequestedRecord, update)
+        if await assert_record_is_new(
+            dao=dao, resource_id=resource_id, update=update, record=record
+        ):
+            await self.delete_file(file_id=resource_id)
+            await dao.insert(record)
 
     async def upsert_file_upload_validation_success(
         self, *, resource_id: str, update: event_schemas.FileUploadValidationSuccess
@@ -341,21 +364,30 @@ class FileRegistry(FileRegistryPort):
             update:
                 The FileUploadValidationSuccess event to upsert.
         """
-        file_without_object_id = models.FileMetadataBase(
-            file_id=update.file_id,
-            decrypted_sha256=update.decrypted_sha256,
-            decrypted_size=update.decrypted_size,
-            upload_date=update.upload_date,
-            decryption_secret_id=update.decryption_secret_id,
-            encrypted_part_size=update.encrypted_part_size,
-            encrypted_parts_md5=update.encrypted_parts_md5,
-            encrypted_parts_sha256=update.encrypted_parts_sha256,
-            content_offset=update.content_offset,
-            storage_alias=update.s3_endpoint_alias,
-        )
+        dao = self._outbox_dao_collection.get_file_upload_validation_success_dao()
 
-        await self.register_file(
-            file_without_object_id=file_without_object_id,
-            staging_object_id=update.object_id,
-            staging_bucket_id=update.bucket_id,
+        record = make_record_from_update(
+            models.FileUploadValidationSuccessRecord, update
         )
+        if await assert_record_is_new(
+            dao=dao, resource_id=resource_id, update=update, record=record
+        ):
+            file_without_object_id = models.FileMetadataBase(
+                file_id=update.file_id,
+                decrypted_sha256=update.decrypted_sha256,
+                decrypted_size=update.decrypted_size,
+                upload_date=update.upload_date,
+                decryption_secret_id=update.decryption_secret_id,
+                encrypted_part_size=update.encrypted_part_size,
+                encrypted_parts_md5=update.encrypted_parts_md5,
+                encrypted_parts_sha256=update.encrypted_parts_sha256,
+                content_offset=update.content_offset,
+                storage_alias=update.s3_endpoint_alias,
+            )
+
+            await self.register_file(
+                file_without_object_id=file_without_object_id,
+                staging_object_id=update.object_id,
+                staging_bucket_id=update.bucket_id,
+            )
+            await dao.insert(record)
